inference_pipeline/requirements.txt
torch==1.9.0
tifffile==2020.9.3
pandas==1.1.5
rasterio==1.2.6
fiona==1.8.22
geopandas==0.9.0
scikit-image
pyyaml